Fixed int to float conversion error.

+++ src/gameclass.cpp
@@ -10,12 +10,12 @@ void Game::Update() {
   cpu.Update(ball.yPos);
   player.Update();
 
-  if (CheckCollisionCircleRec(Vector2{ball.xPos, ball.yPos}, ball.ballRadius, Rectangle{player.xPos, player.yPos, player.width, player.height})) {
+  if (CheckCollisionCircleRec(Vector2{ball.xPos, ball.yPos}, ball.ballRadius, Rectangle{player.xPos, player.yPos, (float)player.width, (float)player.height})) {
     ball.speedX *= -1;
     ball.xPos -= 5;
   }
 
-  if (CheckCollisionCircleRec(Vector2{ball.xPos, ball.yPos}, ball.ballRadius, Rectangle{cpu.xPos, cpu.yPos, cpu.width, cpu.height})) {
+  if (CheckCollisionCircleRec(Vector2{ball.xPos, ball.yPos}, ball.ballRadius, Rectangle{cpu.xPos, cpu.yPos, (float)cpu.width, (float)cpu.height})) {
     ball.speedX *= -1;
     ball.xPos += 5;
   }
